+++ modelopt/__init__.py
@@ -15,6 +15,7 @@
 
 """Nvidia Model Optimizer (modelopt)."""
 
+import warnings as _warnings
 from importlib.metadata import version as _version
 
 __version__ = _version("nvidia-modelopt")